requirements.txt
amqp==5.1.1
APScheduler==3.9.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
cryptography==38.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1
django-allauth==0.51.0
django-apscheduler==0.6.2
django-braces==1.15.0
django-filter==22.1
django-oauth-toolkit==2.1.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.14.0
idna==3.4
jwcrypto==1.4.2
kombu==5.2.4
oauthlib==3.2.1
packaging==21.3
prompt-toolkit==3.0.31
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
types-cryptography==3.3.23
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
